circuits: mpc-circuits: match: Refactor `match` and add tests
