+++ workers/handshake-manager/src/manager/internal_engine.rs
@@ -18,6 +18,7 @@ use rand::{seq::SliceRandom, thread_rng};
 use renegade_crypto::fields::scalar_to_u64;
 use task_driver::settle_match_internal::SettleMatchInternalTask;
 use tracing::log;
+use util::matching_engine::match_orders;
 
 use crate::{
     error::HandshakeManagerError,
@@ -111,14 +112,14 @@ impl HandshakeExecutor {
             // Lookup the other order and match on it
             let res = match self.global_state.get_order(order_id).await {
                 Some(order) => match_orders(
-                    price,
                     my_order,
-                    &my_witness.commitment_witness.balance_send.to_base_type(),
                     &order,
+                    &my_witness.commitment_witness.balance_send.to_base_type(),
                     &other_witness.commitment_witness.balance_send.to_base_type(),
+                    price,
                 ),
                 None => continue,
-            }?;
+            };
 
             // Settle the match if the two orders cross
             if let Some(handshake_result) = res {
@@ -197,375 +198,5 @@ impl HandshakeExecutor {
         .ok_or_else(|| HandshakeManagerError::StateNotFound(ERR_NO_WALLET.to_string()))?;
 
         Ok((order, wallet))
-    }
-}
-
-// -----------
-// | Helpers |
-// -----------
-
-/// Match a given order against all other orders in the wallet
-fn match_orders(
-    midpoint_price: FixedPoint,
-    order1: &Order,
-    balance1: &Balance,
-    order2: &Order,
-    balance2: &Balance,
-) -> Result<Option<MatchResult>, HandshakeManagerError> {
-    // Same asset pair
-    let mut valid_match = order1.base_mint == order2.base_mint
-        && order1.quote_mint == order2.quote_mint
-        && order1.side != order2.side;
-
-    // Validate that the midpoint price is acceptable for both orders
-    valid_match = valid_match
-        && order1.price_in_range(midpoint_price)
-        && order2.price_in_range(midpoint_price);
-
-    // Neither order is zero'd out
-    // Orders are not removed when they are zero'd because a counterparty cannot
-    // update shares to remove an order without revealing the volume of the
-    // order. So zero'd orders may exist in the book until the user removes them
-    valid_match = valid_match && !order1.is_zero() && !order2.is_zero();
-
-    // Compute the amount matched by the engine
-    let party0_max_amount = compute_max_amount(&midpoint_price, order1, balance1);
-    let party1_max_amount = compute_max_amount(&midpoint_price, order2, balance2);
-
-    let min_base_amount = u64::min(party0_max_amount, party1_max_amount);
-    valid_match = valid_match && min_base_amount > 0;
-
-    if !valid_match {
-        return Ok(None);
-    }
-
-    // Compute the auxiliary data for the match
-    let quote_amount = midpoint_price * Scalar::from(min_base_amount);
-    let quote_amount = scalar_to_u64(&quote_amount.floor());
-    let max_minus_min_amount = u64::max(party0_max_amount, party1_max_amount) - min_base_amount;
-
-    Ok(Some(MatchResult {
-        base_mint: order1.base_mint.clone(),
-        quote_mint: order1.quote_mint.clone(),
-        base_amount: min_base_amount,
-        quote_amount,
-        direction: order1.side.into(),
-        max_minus_min_amount,
-        min_amount_order_index: if party0_max_amount <= party1_max_amount {
-            0
-        } else {
-            1
-        },
-    }))
-}
-
-/// Compute the maximum matchable amount for an order and balance
-fn compute_max_amount(price: &FixedPoint, order: &Order, balance: &Balance) -> u64 {
-    match order.side {
-        // Buy the base, the max amount is possibly limited by the quote
-        // balance
-        OrderSide::Buy => {
-            let price_f64 = price.to_f64();
-            let balance_limit = (balance.amount as f64 / price_f64).floor() as u64;
-            u64::min(order.amount, balance_limit)
-        },
-        // Buy the quote, sell the base, the maximum amount is directly limited
-        // by the balance
-        OrderSide::Sell => u64::min(order.amount, balance.amount),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use circuit_types::{
-        balance::Balance,
-        order::{Order, OrderSide},
-    };
-    use lazy_static::lazy_static;
-
-    use crate::manager::internal_engine::match_orders;
-
-    // --------------
-    // | Dummy Data |
-    // --------------
-
-    /// The worst case price for the buy side
-    const BUY_SIDE_WORST_CASE_PRICE: f32 = 10.;
-    /// The worst case price for the sell side
-    const SELL_SIDE_WORST_CASE_PRICE: f32 = 5.;
-
-    lazy_static! {
-        /// The first dummy order used in a valid match
-        static ref ORDER1: Order = Order {
-            base_mint: 1u64.into(),
-            quote_mint: 2u64.into(),
-            side: OrderSide::Buy,
-            amount: 50,
-            worst_case_price: BUY_SIDE_WORST_CASE_PRICE.into(),
-            timestamp: 0,
-        };
-
-        /// The first dummy balance used in a valid match
-        static ref BALANCE1: Balance = Balance {
-            mint: 2u64.into(),
-            amount: 500u64,
-        };
-
-        /// The second dummy order used in a valid match
-        static ref ORDER2: Order = Order {
-            base_mint: 1u64.into(),
-            quote_mint: 2u64.into(),
-            side: OrderSide::Sell,
-            amount: 100,
-            worst_case_price: SELL_SIDE_WORST_CASE_PRICE.into(),
-            timestamp: 0,
-        };
-
-        /// The second dummy balance used in a valid match
-        static ref BALANCE2: Balance = Balance {
-            mint: 1u64.into(),
-            amount: 100u64,
-        };
-    }
-
-    // ---------
-    // | Tests |
-    // ---------
-
-    /// Test a valid match between two orders
-    #[test]
-    fn test_valid_match() {
-        let order1 = ORDER1.clone();
-        let balance1 = BALANCE1.clone();
-        let order2 = ORDER2.clone();
-        let balance2 = BALANCE2.clone();
-        let midpoint_price = 7.;
-
-        let res = match_orders(
-            midpoint_price.into(),
-            &order1,
-            &balance1,
-            &order2,
-            &balance2,
-        )
-        .unwrap();
-
-        assert!(res.is_some());
-        let res = res.unwrap();
-
-        assert_eq!(res.base_mint, 1u64.into());
-        assert_eq!(res.quote_mint, 2u64.into());
-        assert_eq!(res.base_amount, 50);
-        assert_eq!(
-            res.quote_amount,
-            350 // midpoint_price * base_amount
-        );
-        assert_eq!(res.direction, 0);
-        assert_eq!(res.max_minus_min_amount, 50);
-        assert_eq!(res.min_amount_order_index, 0);
-    }
-
-    /// Test a valid match between two order where the buy side is
-    /// undercapitalized
-    #[test]
-    fn test_valid_match_undercapitalized_buy() {
-        let order1 = ORDER1.clone();
-        let mut balance1 = BALANCE1.clone();
-        let order2 = ORDER2.clone();
-        let balance2 = BALANCE2.clone();
-        let midpoint_price = 7.;
-
-        // Can only buy 10 units of the base
-        balance1.amount = (midpoint_price * 10.) as u64;
-
-        let res = match_orders(
-            midpoint_price.into(),
-            &order1,
-            &balance1,
-            &order2,
-            &balance2,
-        )
-        .unwrap();
-
-        assert!(res.is_some());
-        let res = res.unwrap();
-
-        assert_eq!(res.base_mint, 1u64.into());
-        assert_eq!(res.quote_mint, 2u64.into());
-        assert_eq!(res.base_amount, 10);
-        assert_eq!(res.quote_amount, 70 /* midpoint_price * base_amount */);
-        assert_eq!(res.direction, 0);
-        assert_eq!(res.max_minus_min_amount, 90);
-        assert_eq!(res.min_amount_order_index, 0);
-    }
-
-    /// Test a valid match between two order where the sell side is
-    /// undercapitalized
-    #[test]
-    fn test_valid_match_undercapitalized_sell() {
-        let order1 = ORDER1.clone();
-        let balance1 = BALANCE1.clone();
-        let order2 = ORDER2.clone();
-        let mut balance2 = BALANCE2.clone();
-        let midpoint_price = 7.;
-
-        // Can only sell 10 units of the base
-        balance2.amount = 10u64;
-
-        let res = match_orders(
-            midpoint_price.into(),
-            &order1,
-            &balance1,
-            &order2,
-            &balance2,
-        )
-        .unwrap();
-
-        assert!(res.is_some());
-        let res = res.unwrap();
-
-        assert_eq!(res.base_mint, 1u64.into());
-        assert_eq!(res.quote_mint, 2u64.into());
-        assert_eq!(res.base_amount, 10);
-        assert_eq!(res.quote_amount, 70 /* midpoint_price * base_amount */);
-        assert_eq!(res.direction, 0);
-        assert_eq!(res.max_minus_min_amount, 40);
-        assert_eq!(res.min_amount_order_index, 1);
-    }
-
-    /// Test mismatched base mints
-    #[test]
-    fn test_mismatched_base_mints() {
-        let order1 = ORDER1.clone();
-        let balance1 = BALANCE1.clone();
-        let mut order2 = ORDER2.clone();
-        let balance2 = BALANCE2.clone();
-
-        order2.base_mint = 3u64.into();
-        let midpoint_price = 7.;
-
-        let res = match_orders(
-            midpoint_price.into(),
-            &order1,
-            &balance1,
-            &order2,
-            &balance2,
-        )
-        .unwrap();
-
-        assert!(res.is_none());
-    }
-
-    /// Test mismatched quote mints
-    #[test]
-    fn test_mismatched_quote_mints() {
-        let order1 = ORDER1.clone();
-        let balance1 = BALANCE1.clone();
-        let mut order2 = ORDER2.clone();
-        let balance2 = BALANCE2.clone();
-
-        order2.quote_mint = 3u64.into();
-        let midpoint_price = 7.;
-
-        let res = match_orders(
-            midpoint_price.into(),
-            &order1,
-            &balance1,
-            &order2,
-            &balance2,
-        )
-        .unwrap();
-
-        assert!(res.is_none());
-    }
-
-    /// Test two orders on the same side of the book
-    #[test]
-    fn test_same_side() {
-        let order1 = ORDER1.clone();
-        let balance1 = BALANCE1.clone();
-        let mut order2 = ORDER2.clone();
-        let balance2 = BALANCE2.clone();
-
-        order2.side = order1.side;
-        let midpoint_price = 7.;
-
-        let res = match_orders(
-            midpoint_price.into(),
-            &order1,
-            &balance1,
-            &order2,
-            &balance2,
-        )
-        .unwrap();
-
-        assert!(res.is_none());
-    }
-
-    /// Test a midpoint price out of the buy side range
-    #[test]
-    fn test_buy_price_out_of_range() {
-        let order1 = ORDER1.clone();
-        let balance1 = BALANCE1.clone();
-        let order2 = ORDER2.clone();
-        let balance2 = BALANCE2.clone();
-
-        let midpoint_price = BUY_SIDE_WORST_CASE_PRICE + 1.;
-
-        let res = match_orders(
-            midpoint_price.into(),
-            &order1,
-            &balance1,
-            &order2,
-            &balance2,
-        )
-        .unwrap();
-
-        assert!(res.is_none());
-    }
-
-    /// Test a midpoint price out of the sell side range
-    #[test]
-    fn test_sell_price_out_of_range() {
-        let order1 = ORDER1.clone();
-        let balance1 = BALANCE1.clone();
-        let order2 = ORDER2.clone();
-        let balance2 = BALANCE2.clone();
-
-        let midpoint_price = SELL_SIDE_WORST_CASE_PRICE - 1.;
-
-        let res = match_orders(
-            midpoint_price.into(),
-            &order1,
-            &balance1,
-            &order2,
-            &balance2,
-        )
-        .unwrap();
-
-        assert!(res.is_none());
-    }
-
-    /// Tests a balance of zero on one side
-    #[test]
-    fn test_zero_balance() {
-        let order1 = ORDER1.clone();
-        let balance1 = BALANCE1.clone();
-        let order2 = ORDER2.clone();
-        let mut balance2 = BALANCE2.clone();
-
-        balance2.amount = 0;
-        let midpoint_price = 7.;
-
-        let res = match_orders(
-            midpoint_price.into(),
-            &order1,
-            &balance1,
-            &order2,
-            &balance2,
-        )
-        .unwrap();
-
-        assert!(res.is_none());
     }
 }
